Fix issue where two clients could obtain the same IP
Any time a write happens to either nics::leased::<nic> or ips::leased::<ip>,
it needs to happen in a transaction that ensures that either the contents are
not changing, or the keys are missing. These keys should only ever be deleted
(without administrator action) by a lease timing out, and should never have
their values modified after creation. The only edit should be updating the
etcd leases.